+++ src/ai/AiModelManager.ts
@@ -1,0 +1,29 @@
+import {AiModel} from "./core/AiModel.ts";
+import {AiGlobalConfig} from "./AiGlobalConfig.ts";
+import {SparkAiModel} from "./spark/SparkAiModel.ts";
+import {Editor} from "@tiptap/core";
+
+export class AiModelManager {
+
+    private static models: Record<string, AiModel> = {};
+
+    static init(editor: Editor, globalConfig: AiGlobalConfig) {
+        if (globalConfig && globalConfig.models) {
+            for (let key of Object.keys(globalConfig.models)) {
+                switch (key) {
+                    case "spark":
+                        this.set(key, new SparkAiModel(editor, globalConfig))
+                }
+            }
+        }
+    }
+
+    static get(modelName: string): AiModel {
+        return this.models[modelName];
+    }
+
+    static set(modelName: string, aiModel: AiModel) {
+        this.models[modelName] = aiModel;
+    }
+
+}